requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
pymysql==1.1.0
mysql-connector-python==8.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
celery==5.3.6
redis==5.0.1
selenium==4.17.2
webdriver-manager==4.0.1
pydantic-settings==2.1.0